pyproject.toml
[tool.poetry]
name = "autosticker"
version = "0.1.0"
description = ""
authors = ["Robert Dumitru <72272382+robert-dumitru@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.3.0"
openai = "^0.25.0"
replicate = "^0.4.0"
requests = "^2.28.1"
pyTelegramBotAPI = "^4.7.1"
setuptools = "^65.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
